server/requirements.txt
requests
Django
gunicorn
Pillow
ibm-watson
python-dotenv
